pyproject.toml
[tool.poetry]
name = "notifications"
version = "0.1.0"
description = ""
authors = ["Artem <clasterpack@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.2"
django-rest-framework = "^0.1.0"
celery = "^5.2.7"
drf-yasg = "^1.21.4"
python-dotenv = "^0.21.0"
redis = "^4.3.4"
django-celery-results = "^2.4.0"
flower = "^1.2.0"
amqp = "^5.1.1"
asgiref = "^3.5.2"
pytest = {version = "^7.1.3", optional = true}
isort = {version = "^5.10.1", optional = true}
django-prometheus = "^2.2.0"
freezegun = {version = "^1.2.2", optional = true}
flake8 = {version = "4.0.1", optional = true}
wemake-python-styleguide = {version = "^0.17.0", optional = true}


[tool.poetry.extras]
devtools = ["pytest", "isort", "wemake-python-styleguide", "freezegun", "flake8",]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
